fix(BUG): delete alert rule is fixed
